Move index generator to the end of array
When "index" generator is set as the first one, there is a problem
with generating index page for other than default language.
Suggesting this order in official documentation may lead to
frustration, because solution is not so obvious and most of the
new users follow the instruction by copy/paste method.
This lead to situation where generator not working as it should.

Let's inform about "valid" order from the beginning.